+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Remove GHC uncompressed tar file after installation [#376](https://github.com/commercialhaskell/stack/issues/376)
+* Put stackage snapshots JSON on S3 [#380](https://github.com/commercialhaskell/stack/issues/380)
 
 ## 0.1.0.0