+++ ChangeLog.md
@@ -6,6 +6,7 @@
 * single test suite failure should show entire log [#388](https://github.com/commercialhaskell/stack/issues/388)
 * valid-wanted is a confusing option name [#386](https://github.com/commercialhaskell/stack/issues/386)
 * stack init in multi-package project should use local packages for dependency checking [#384](https://github.com/commercialhaskell/stack/issues/384)
+* Display information on why a snapshot was rejected [#381](https://github.com/commercialhaskell/stack/issues/381)
 
 ## 0.1.0.0