+++ ChangeLog.md
@@ -7,6 +7,7 @@
 * valid-wanted is a confusing option name [#386](https://github.com/commercialhaskell/stack/issues/386)
 * stack init in multi-package project should use local packages for dependency checking [#384](https://github.com/commercialhaskell/stack/issues/384)
 * Display information on why a snapshot was rejected [#381](https://github.com/commercialhaskell/stack/issues/381)
+* Give a reason for unregistering packages [#389](https://github.com/commercialhaskell/stack/issues/389)
 
 ## 0.1.0.0