+++ ChangeLog.md
@@ -10,6 +10,7 @@
 * Give a reason for unregistering packages [#389](https://github.com/commercialhaskell/stack/issues/389)
 * `stack exec` accepts the `--no-ghc-package-path` parameter
 * Don't require build plan to upload [#400](https://github.com/commercialhaskell/stack/issues/400)
+* Specifying test components only builds/runs those tests [#398](https://github.com/commercialhaskell/stack/issues/398)
 
 Bug fixes: