+++ ChangeLog.md
@@ -17,6 +17,7 @@ Bug fixes:
 
 * Haddock links to global packages no longer broken on Windows [#375](https://github.com/commercialhaskell/stack/issues/375)
 * Make flags case-insensitive [#397](https://github.com/commercialhaskell/stack/issues/397)
+* Mark packages uninstalled before rebuilding [#365](https://github.com/commercialhaskell/stack/issues/365)
 
 ## 0.1.0.0