+++ ChangeLog.md
@@ -13,6 +13,7 @@
 * Specifying test components only builds/runs those tests [#398](https://github.com/commercialhaskell/stack/issues/398)
 * `STACK_EXE` environment variable
 * Add the `stack dot` command
+* `stack upgrade` added [#237](https://github.com/commercialhaskell/stack/issues/237)
 
 Bug fixes: