+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.1.1.0
 
 * Remove GHC uncompressed tar file after installation [#376](https://github.com/commercialhaskell/stack/issues/376)
 * Put stackage snapshots JSON on S3 [#380](https://github.com/commercialhaskell/stack/issues/380)